pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "idse-orchestrator"
version = "0.1.0"
description = "CLI tool for managing Intent-Driven Systems Engineering (IDSE) projects"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "IDSE Developer Agency", email = "noreply@idse-agency.dev"}
]
keywords = ["idse", "development", "pipeline", "orchestration", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
idse = "idse_orchestrator.cli:main"

[project.urls]
Homepage = "https://github.com/tjpilant/idse-orchestrator"
Documentation = "https://github.com/tjpilant/idse-orchestrator#readme"
Repository = "https://github.com/tjpilant/idse-orchestrator"
